ENH data type changes

+++ cpp/include/cuml/explainer/kernel_shap.hpp
@@ -40,13 +40,13 @@ namespace Explainer{
  * @param[in]    seed               Seed for the random number generator
  * @{
  */
-void kernel_dataset(const raft::handle_t& handle, int* X, int nrows_X,
+void kernel_dataset(const raft::handle_t& handle, float* X, int nrows_X,
                     int M, float* background, int nrows_background,
                     float* combinations, float* observation,
                     int* nsamples, int len_nsamples, int maxsample,
                     uint64_t seed = 0ULL);
 
-void kernel_dataset(const raft::handle_t& handle, int* X, int nrows_X,
+void kernel_dataset(const raft::handle_t& handle, double* X, int nrows_X,
                     int M, double* background, int nrows_background,
                     double* combinations, double* observation,
                     int* nsamples, int len_nsamples, int maxsample,
